Mode back to triestore for review
